Streamlined the copy to the staging directory in upload_directory.
For symlinks, we just reproduce them exactly, regardless of whether they are
local or not. If they're non-local links that end up being dangling after the
copy... well too bad, the user shouldn't have created them.

For copies, we don't bother creating hardlinks. This is safer as it avoids
propagation of hardlinks into the registry, whereby users could then
inadvertently modify our "immutable" files. Performance should not be majorly
impacted as we were copying across partitions anyway.

To offset any perf hit, the process of copying the files is now parallelized.
We also check that the copied files have the correct MD5 checksums and sizes,
just to protect against filesystem corruption.

Finally, we reproduce the empty directories in the empty directory.